began moving the cardflip.js logic into Card.tsx but i dont think its quite working. not tested. realized i need to do all of this for SpyCard.tsx as well
